Ahora la rejilla guarda los datos del terreno y del agua por separado, además de guardarlos como double.
git-svn-id: https://forja.rediris.es/svn/cusl4-catastrof@19 8c4acde5-3c00-4ed3-8754-bc08a9928e99

+++ agents/WaterAgent.java
@@ -29,13 +29,13 @@ public class WaterAgent extends Agent {
 		Object[] args = getArguments();
 		int x; // Posición inicial en la rejilla
 		int y;
-		int value; // Potencial de la casilla inicial
-		int water; // Cantidad de agua del agente
+		double value; // Potencial de la casilla inicial
+		double water; // Cantidad de agua del agente
 		if (args != null && args.length == 4) {
 			x = Integer.parseInt((String) args[0]);
 			y = Integer.parseInt((String) args[1]);
-			value = Integer.parseInt((String) args[2]);
-			water = Integer.parseInt((String) args[3]);
+			value = Double.parseDouble((String) args[2]);
+			water = Double.parseDouble((String) args[3]);
 		} else {
 			throw new IllegalArgumentException("Wrong arguments.");
 		}
